Health_Bloom/requirements.txt
pandas==1.3.2
Werkzeug==2.0.1
keras==2.6.0
Flask_Bootstrap==3.3.7.1
Flask_SQLAlchemy==2.5.1
numpy==1.21.0
Flask_WTF==1.0.0
Flask_Login==0.5.0
WTForms==3.0.0
Flask==2.0.2
opencv-python-headless
Flask_Bcrypt==0.7.1
Pillow==9.0.1
scikit_learn==1.0.2
